Refactor brain helpers into typed modules
